Test updates.

+++ .travis/test.sh
@@ -4,5 +4,10 @@ set -e
 PYTEST_OPTIONS=(--cov=${MODULE} -xvs --pyargs ${MODULE})
 COVERAGE_FILE=${TEST_DIR}/.coverage
 
+# ensure a clean temporary test directory
+[[ -d ${TEST_DIR} ]] && rm -rf ${TEST_DIR}
+cd ${TEST_DIR}
+
 py.test "${PYTEST_OPTIONS[@]}" .
-cp ${COVERAGE_FILE} ${TRAVIS_BUILD_DIR}
+[[ -f ${COVERAGE_FILE} ]] && cp ${COVERAGE_FILE} ${TRAVIS_BUILD_DIR} || echo "Warning: code coverage file not available."
+cd ${OLDPWD}
